Teams
During this pandemic, we all are compelled to attend online classes and I’m sure most of you are worried about your attendance during online classes. Sometimes you might not be in a position to attend it and hence you will lose your attendance for the day. Isn’t that problematic?
So i have come up with a solution to it! You can make a bot that will join/attend your classes for you! Yes, you’ve heard it right! Now you can get attendance from wherever you are and whatever you are doing. Wanna know how? For the tutorial, Click on the link given below-
https://www.youtube.com/watch?v=WfPk86Qf1Lo